migrate directional-stylesheet utility
