+++ packages/web-components/fast-components/src/switch/switch.styles.ts
@@ -1,6 +1,5 @@
 import { css } from "@microsoft/fast-element";
 import {
-    DirectionalStyleSheetBehavior,
     disabledCursor,
     display,
     focusVisible,
@@ -28,7 +27,7 @@ import {
     typeRampBaseFontSize,
     typeRampBaseLineHeight,
 } from "../design-tokens";
-import { heightNumber } from "../styles/index";
+import { DirectionalStyleSheetBehavior, heightNumber } from "../styles/index";
 
 export const switchStyles = (context, definition) =>
     css`
