+++ packages/web-components/fast-foundation/src/utilities/style/index.ts
@@ -1,4 +1,3 @@
 export * from "./disabled";
 export * from "./display";
 export * from "./focus";
-export * from "./direction";
